Why are our best and most experienced employees leaving prematurely?

A dataset from Kaggle competition Human Resources Analytics. https://www.kaggle.com/

Format

A data frame with 14999 rows and 10 variables

Source

Dataset HR-analytics from https://www.kaggle.com

Details

satisfaction_level Level of satisfaction (0-1)

last_evaluation Time since last performance evaluation (in Years)

number_project Number of projects completed while at work

average_montly_hours Average monthly hours at workplace

time_spend_company Number of years spent in the company

Work_accident Whether the employee had a workplace accident

left Whether the employee left the workplace or not (1 or 0) Factor

promotion_last_5years Whether the employee was promoted in the last five years

sales Department in which they work for

salary Relative level of salary (high)